+++ python/PDFs/physics/CMakeLists.txt
@@ -2,6 +2,7 @@ goofit_add_library(_Physics STATIC
     AmpNBodyBase.cpp
     Amp3BodyBase.cpp
     Amp3Body.cpp
+    Amp3BodyT.cpp
     DalitzVetoPdf.cpp
     DalitzPlotHelpers.cpp
     DalitzPlotter.cpp
@@ -12,7 +13,6 @@ goofit_add_library(_Physics STATIC
     ResonancePdf.cpp
     SpinFactors.cpp
     Tddp4Pdf.cu
-    TddpPdf.cpp
     ThreeGaussResolution_Aux.cpp
     TruthResolution_Aux.cpp
 )
@@ -24,6 +24,6 @@ goofit_add_python_docs(_Physics PDFs/physics/DalitzVetoPdf.h)
 goofit_add_python_docs(_Physics PDFs/physics/IncoherentSumPdf.h)
 goofit_add_python_docs(_Physics PDFs/physics/MixingTimeResolution_Aux.h)
 goofit_add_python_docs(_Physics PDFs/physics/ResonancePdf.h)
-goofit_add_python_docs(_Physics PDFs/physics/TddpPdf.h)
+goofit_add_python_docs(_Physics PDFs/physics/Amp3BodyT.h)
 goofit_add_python_docs(_Physics PDFs/physics/ThreeGaussResolution_Aux.h)
 goofit_add_python_docs(_Physics PDFs/physics/TruthResolution_Aux.h)
