+++ python/goofit.cpp
@@ -52,6 +52,7 @@ void init_DalitzPlotHelpers(py::module &);
 void init_AmpNBodyBase(py::module &);
 void init_Amp3BodyBase(py::module &);
 void init_Amp3Body(py::module &);
+void init_Amp3BodyT(py::module &);
 void init_DalitzPlotter(py::module &);
 void init_DalitzVetoPdf(py::module &);
 void init_DP4Pdf(py::module &);
@@ -61,7 +62,6 @@ void init_MixingTimeResolution(py::module &);
 void init_ResonancePdf(py::module &);
 void init_SpinFactors(py::module &);
 void init_Tddp4Pdf(py::module &);
-void init_TddpPdf(py::module &);
 void init_ThreeGaussResolution(py::module &);
 void init_TruthResolution(py::module &);
 
@@ -122,6 +122,7 @@ PYBIND11_MODULE(_goofit, m) {
     init_AmpNBodyBase(m);
     init_Amp3BodyBase(m);
     init_Amp3Body(m);
+    init_Amp3BodyT(m);
     init_DalitzVetoPdf(m);
     init_DP4Pdf(m);
     init_IncoherentSumPdf(m);
@@ -130,7 +131,6 @@ PYBIND11_MODULE(_goofit, m) {
     init_ResonancePdf(m);
     init_SpinFactors(m);
     init_Tddp4Pdf(m);
-    init_TddpPdf(m);
     init_ThreeGaussResolution(m);
     init_TruthResolution(m);
